Fix licenses
