+++ ee/vault/vault.go
@@ -1,7 +1,7 @@
 // +build oss
 
 /*
- * Copyright 2021 Dgraph Labs, Inc. and Contributors
+ * Copyright 2020-2021 Dgraph Labs, Inc. and Contributors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,9 +20,11 @@ package vault
 
 import (
 	"github.com/dgraph-io/dgraph/x"
+	"github.com/golang/glog"
 	"github.com/spf13/viper"
 )
 
 func GetKeys(config *viper.Viper) (aclKey, encKey x.SensitiveByteSlice) {
 	glog.Fatalf("vault: this is an enterprise-only feature")
+	return
 }
